+++ src/plugins/timeline/systems/layout_bundles.rs
@@ -38,7 +38,7 @@ pub fn create_empty_grid_area() -> NodeBundle {
   }
 }
 
-pub fn create_top_bar_grid_area() -> NodeBundle {
+pub fn create_topbar_grid_area() -> NodeBundle {
   NodeBundle {
     style: Style {
       display: Display::Grid,
@@ -91,68 +91,3 @@ pub fn create_side_panel_grid_area() -> NodeBundle {
   }
 }
 
-pub fn create_timeline_col_container() -> NodeBundle {
-  NodeBundle {
-    style: Style {
-      display: Display::Flex,
-      flex_direction: FlexDirection::Column,
-      align_items: AlignItems::Center,
-      justify_content: JustifyContent::Center,
-      column_gap: Val::Px(3.),
-      width: Val::Percent(100.),
-      height: Val::Percent(100.),
-      padding: UiRect::all(Val::Px(5.0)).with_bottom(Val::Px(10.)),
-      ..default()
-    },
-    background_color: BackgroundColor(Color::BLACK),
-    border_radius: BorderRadius::px(0., 0., 0., 16.),
-    ..default()
-  }
-}
-
-pub fn create_timeline_hours_row_container() -> NodeBundle {
-  NodeBundle {
-    style: Style {
-      display: Display::Flex,
-      flex_direction: FlexDirection::Row,
-      align_items: AlignItems::Center,
-      justify_content: JustifyContent::SpaceEvenly,
-      width: Val::Percent(100.),
-      height: Val::Px(28.),
-      ..default()
-    },
-    background_color: BackgroundColor(Color::BLACK),
-    border_radius: BorderRadius::px(0., 0., 0., 16.),
-    ..default()
-  }
-}
-
-pub fn create_timeline_hour_outer() -> NodeBundle {
-  NodeBundle {
-    style: Style {
-      display: Display::Block,
-      width: Val::Percent(3.),
-      height: Val::Px(24.),
-      padding: UiRect { top: Val::Px(2.), bottom: Val::Px(0.), ..default() },
-      ..default()
-    },
-    background_color: BackgroundColor(Color::WHITE),
-    ..default()
-  }
-}
-
-pub fn create_timeline_hour_inner() -> NodeBundle {
-  NodeBundle {
-    style: Style {
-      display: Display::Flex,
-      align_items: AlignItems::Center,
-      justify_content: JustifyContent::Center,
-      width: Val::Percent(100.),
-      height: Val::Px(22.),
-      padding: UiRect { top: Val::Px(4.), ..default() },
-      ..default()
-    },
-    background_color: BackgroundColor(Color::BLACK),
-    ..default()
-  }
-}
